+++ M.Resources/Localisation/LLin/BaseStrings.zh.resx
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<root>
+    <xsd:schema
+            xmlns:xsd="http://www.w3.org/2001/XMLSchema"
+            xmlns:msdata="urn:schemas-microsoft-com:xml-msdata" id="root">
+        <xsd:element name="root" msdata:IsDataSet="true" />
+    </xsd:schema>
+    <resheader name="resmimetype">
+        <value>text/microsoft-resx</value>
+    </resheader>
+    <resheader name="version">
+        <value>1.3</value>
+    </resheader>
+    <resheader name="reader">
+        <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=2.0.0.0, Culture=neutral,
+            PublicKeyToken=b77a5c561934e089</value>
+    </resheader>
+    <resheader name="writer">
+        <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=2.0.0.0, Culture=neutral,
+            PublicKeyToken=b77a5c561934e089</value>
+    </resheader>
+    <data name="exit" xml:space="preserve">
+        <value>退出</value>
+    </data>
+    <data name="manual" xml:space="preserve">
+        <value>食用手册</value>
+    </data>
+    <data name="prev_restart" xml:space="preserve">
+        <value>上一首 / 重新开始</value>
+    </data>
+    <data name="next" xml:space="preserve">
+        <value>下一首</value>
+    </data>
+    <data name="toggle_pause" xml:space="preserve">
+        <value>暂停 / 播放</value>
+    </data>
+    <data name="view_plugins" xml:space="preserve">
+        <value>查看插件</value>
+    </data>
+    <data name="hide_and_lock_interface" xml:space="preserve">
+        <value>锁定变更并隐藏界面</value>
+    </data>
+    <data name="lock_interface" xml:space="preserve">
+        <value>锁定变更</value>
+    </data>
+    <data name="toggle_loop" xml:space="preserve">
+        <value>单曲循环</value>
+    </data>
+    <data name="view_in_song_select" xml:space="preserve">
+        <value>在单人游戏选歌界面中查看</value>
+    </data>
+    <data name="open_sidebar" xml:space="preserve">
+        <value>侧边栏</value>
+    </data>
+    <data name="audio_control_requested_main" xml:space="preserve">
+        <value>请求接手音频控制。</value>
+    </data>
+    <data name="audio_control_requested_sub" xml:space="preserve">
+        <value>原因是: {0}</value>
+    </data>
+</root>
