+++ osu.Game.Rulesets.IGPlayer/Feature/Player/Plugins/Bundle/Collection/Sidebar/CollectionFunctionProvider.cs
@@ -1,7 +1,7 @@
-using M.Resources.Localisation.LLin.Plugins;
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Rulesets.IGPlayer.Feature.Player.Graphics;
 using osu.Game.Rulesets.IGPlayer.Feature.Player.Plugins.Types;
+using osu.Game.Rulesets.IGPlayer.Localisation.LLin.Plugins;
 
 namespace osu.Game.Rulesets.IGPlayer.Feature.Player.Plugins.Bundle.Collection.Sidebar
 {
