+++ osu.Game.Rulesets.IGPlayer/Helper/Injectors/PreviewTrackInjector.cs
@@ -4,24 +4,17 @@ using osu.Framework.Allocation;
 using osu.Framework.Audio;
 using osu.Framework.Bindables;
 using osu.Framework.Development;
-using osu.Framework.Graphics;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Logging;
-using osu.Framework.Utils;
 using osu.Game.Audio;
 using osu.Game.Beatmaps;
 using osu.Game.Database;
-using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterface;
 using osu.Game.Online.API;
 using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Overlays;
 using osu.Game.Rulesets.IGPlayer.Feature.DownloadAccel;
 using osu.Game.Rulesets.IGPlayer.Feature.DownloadAccel.Graphics;
-using osu.Game.Tests.Visual;
-using osuTK;
 using Realms;
 
 namespace osu.Game.Rulesets.IGPlayer.Helper.Injectors;
