+++ osu.Game.Rulesets.IGPlayer/Localisation/LLin/Plugins/YaspStrings.cs
@@ -1,0 +1,13 @@
+using osu.Framework.Localisation;
+
+namespace osu.Game.Rulesets.IGPlayer.Localisation.LLin.Plugins
+{
+    public static class YaspStrings
+    {
+        private const string prefix = @"M.Resources.Localisation.LLin.Plugins.YaspStrings";
+
+        public static LocalisableString Scale => new TranslatableString(getKey(@"scale"), @"经典样式缩放");
+
+        private static string getKey(string key) => $@"{prefix}:{key}";
+    }
+}
