+++ osu.Game.Rulesets.IGPlayer/osu.Game.Rulesets.IGPlayer.csproj
@@ -27,7 +27,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Helper" />
-    <Folder Include="Resources" />
     <Folder Include="Rs" />
   </ItemGroup>
 </Project>
